requirements.txt
bert4keras==0.11.3
Keras==2.3.1
tensorflow==1.15.4
numpy==1.18.5
PyYAML==6.0
Flask==2.0.3
tqdm==4.62.3